+++ Minigin/Move.cpp
@@ -102,8 +102,8 @@ namespace dae
 
 	bool Move::CanGoRight(int idx)
 	{
-		int rightIdx{ (idx + 1) };
-		if (rightIdx % m_GridHeight <= m_GridWidth && idx + 1 > 0 && m_LastTileIdx != rightIdx && rightIdx < m_Grid.size())
+		size_t rightIdx{ static_cast<size_t>(idx + 1) };
+		if (static_cast<int>(rightIdx) % m_GridHeight <= m_GridWidth && idx + 1 > 0 && m_LastTileIdx != static_cast<int>(rightIdx) && rightIdx < m_Grid.size())
 		{
 			if (m_Grid[rightIdx].walkable)
 				return true;
@@ -127,8 +127,8 @@ namespace dae
 
 	bool Move::CanGoDown(int idx)
 	{
-		int downIdx{ idx + m_GridWidth };
-		if (downIdx <= m_GridWidth * m_GridHeight && idx + m_GridWidth > 0 && m_LastTileIdx != downIdx && downIdx < m_Grid.size())
+		size_t downIdx{ static_cast<size_t>(idx + m_GridWidth) };
+		if (static_cast<int>(downIdx) <= m_GridWidth * m_GridHeight && idx + m_GridWidth > 0 && m_LastTileIdx != static_cast<int>(downIdx) && downIdx < m_Grid.size())
 		{
 			if (m_Grid[downIdx].walkable)
 				return true;
